Refactor performance counter imports and consolidate helper functions into _lib_helpers.py
